+++ tests/RocketChatWebhookChannelTest.php
@@ -11,6 +11,7 @@ use GuzzleHttp\Client;
 use GuzzleHttp\Psr7\Response;
 use Illuminate\Notifications\Notification;
 use NotificationChannels\RocketChat\RocketChat;
+use NotificationChannels\RocketChat\RocketChatAttachment;
 use NotificationChannels\RocketChat\RocketChatMessage;
 use NotificationChannels\RocketChat\RocketChatWebhookChannel;
 use NotificationChannels\RocketChat\Exceptions\CouldNotSendNotification;
@@ -71,6 +72,7 @@ class RocketChatWebhookChannelTest extends TestCase
         $channel->send(new TestNotifiable(), new TestNotification());
     }
 
+
     /** @test */
     public function it_does_not_send_a_message_when_room_missed()
     {
@@ -107,6 +109,18 @@ class TestNotification extends Notification
     public function toRocketChat()
     {
         return RocketChatMessage::create('hello')->from(':token')->to(':room');
+    }
+}
+class TestNotificationWithContent extends Notification
+{
+    public function toRocketChat()
+    {
+        return RocketChatMessage::create('hello')
+            ->alias('Botty Bot')
+            ->attachment(
+                RocketChatAttachment::create()
+                ->imageUrl('https://picsum.photos/200/300')
+            );
     }
 }