+++ CHANGELOG.md
@@ -5,3 +5,6 @@ All notable changes to `laravel-notification-channels/rocket-chat` will be docum
 ## 0.1.0 - 2017-03-02
 
 - Initial release
+
+## 2020-02-20
+- adding attachments to messages